Added exception not thrown tests in encapsulate all other tests
It is possible for tests to throw exceptions, but rather than creating a usable
Failure report to investigate what is failing, node_unit reports the actual
exceptions which are very unhelpful.  By encapsulating all tests in a
test.DoesNotThrow() block, exceptions are caught and reported along with all
other failing tests. One caveat is that if a exception is caught, subsequent
tests are skipped causing the expected number of tests test to fail.