Every restclient test setup does a good login, except for authentication tests
To improve stability and usability of the rest client tests but at the expense
of speed, every test case except for authentication tests should perform an good
login on start up. This is favorable to depending on the last test to adequately
set the api-token for the next restclient call used for the next test. This also
keeps tests isolated from each other by giving each test case a separate login
session.